megano/requirements.txt
asgiref==3.7.2
attrs==23.1.0
diploma-frontend @ file:///home/ivan/Desktop/SkillboxDiplom/diploma-frontend/dist/diploma-frontend-0.6.tar.gz
Django==4.2.1
django-filter==23.2
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
inflection==0.5.1
jsonschema==4.17.3
Pillow==9.5.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
